change insert request timeout implementation to enable sharing http client instance
This change introducing using the context.Context member on the
http.Request struct for timeouts rather than creating a new http.Client
for every request. This allows for connection to be cached (so that a
users are not forced to establish a new https connection for every
request) while maintaining the same functionality. A test was also
added to prove that timeouts still work.