add Statistics.ProcessedEventCount
The goal of ProcessedEventCount is to enable users to know if all
events that have been enqueued have been sent to insights (or given
up on because of too many failed tries). This should aid in being
able to shutdown more efficiently and still ensure all events have
been flushed.

This also adds documentation to the Statistics struct.