add EnqueueEventContext to help prevent potential of blocking
Though it hopefully wouldn't be reality long term in production, there is
the potential that for very high volume and too few workers there would be
the potential for blocking on putting on the channel for the workers to
pick up. This allows the user of the client to decide how long they would
like to block before giving up trying to put on the channel.